feat: adds alerts to authoring interfaces
